move: sizeInput 이동
